% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{align_rasters}
\alias{align_rasters}
\title{Align Raster Extents}
\usage{
align_rasters(r1, r2, verbose = TRUE)
}
\arguments{
\item{r1}{First raster (RasterLayer)}

\item{r2}{Second raster (RasterLayer)}

\item{verbose}{Logical. Print progress messages?}
}
\value{
List with two cropped rasters (named 'before' and 'after')
}
\description{
Crops rasters to their common spatial extent.
}
\keyword{internal}
