% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diff.R
\name{calculate_difference}
\alias{calculate_difference}
\title{Calculate Pixel-wise Differences}
\usage{
calculate_difference(
  before_raster,
  after_raster,
  method = c("simple", "absolute", "normalized")
)
}
\arguments{
\item{before_raster, after_raster}{RasterLayer objects}

\item{method}{"simple", "absolute", or "normalized"}
}
\value{
Difference raster
}
\description{
Calculate Pixel-wise Differences
}
