% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{calculate_threshold}
\alias{calculate_threshold}
\title{Automatic threshold calculation}
\usage{
calculate_threshold(r, method = c("otsu", "kmeans"))
}
\arguments{
\item{r}{RasterLayer}

\item{method}{Threshold method}
}
\description{
Automatic threshold calculation
}
