% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_chng.R
\name{detect_chng}
\alias{detect_chng}
\title{Detect Significant Changes Using Thresholding}
\usage{
detect_chng(
  diff_raster,
  threshold = NULL,
  method = c("fixed", "otsu", "kmeans")
)
}
\arguments{
\item{diff_raster}{RasterLayer of differences}

\item{threshold}{Numeric threshold value}

\item{method}{Threshold method ("fixed", "otsu", "kmeans")}
}
\value{
Binary RasterLayer (1 = change, 0 = no change)
}
\description{
Detect Significant Changes Using Thresholding
}
