% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{normalize_difference}
\alias{normalize_difference}
\title{Calculate normalized difference}
\usage{
normalize_difference(r1, r2)
}
\arguments{
\item{r1, r2}{RasterLayers}
}
\value{
Normalized difference raster
}
\description{
Calculate normalized difference
}
