% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_rastdat.R
\name{prep_rastdat}
\alias{prep_rastdat}
\title{Preprocess Raster Data for Change Detection}
\usage{
prep_rastdat(
  before_raster,
  after_raster,
  target_resolution = NULL,
  resample_method = "bilinear",
  na_action = "replace",
  na_value = 0,
  verbose = TRUE
)
}
\arguments{
\item{before_raster, }{after_raster RasterLayer objects}

\item{after_raster}{A RasterLayer or SpatRaster representing the "after" time period.}

\item{target_resolution}{Numeric vector (xres, yres) or NULL}

\item{resample_method}{"bilinear" or "ngb"}

\item{na_action}{"replace" or "ignore"}

\item{na_value}{Replacement value for NAs}

\item{verbose}{Print progress messages?}
}
\value{
List of processed rasters
}
\description{
Preprocess Raster Data for Change Detection
}
